Add points for approval
